% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.values.R
\name{fitted.values}
\alias{fitted.values}
\title{Fitted Values for Theta on Beta Regression}
\usage{
fitted.values(x)
}
\arguments{
\item{x}{an object of the class \emph{bayesbr}, containing the list returned from the \code{\link{bayesbr}} function.}
}
\value{
A vector with the average of theta estimates in the iterations (excluding warmup). The vector size is equal to the number of model observations.
}
\description{
A function that receives information from an estimated model uses data from the estimated theta for each iteration and returns the average of each theta in the sample.
}
\seealso{
\code{\link{bayesbr}},\code{\link{predict.bayesbr}}
}
