% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_plot.R
\name{loglikPlot}
\alias{loglikPlot}
\title{Plot Chain of Loglik Using GGplot}
\usage{
loglikPlot(loglik)
}
\arguments{
\item{loglik}{A vector with the estimated model loglik chain}
}
\description{
The function receives a vector containing the model's loglik chain and displays it through a graph using the GGplot package, through this graph it is possible to see if the model has converged.
}
\examples{
data("CarTask", package = "bayesbr")

bbr = bayesbr(probability~task + NFCCscale,data = CarTask,
             iter = 100, mean_betas = c(1, 0.5,1.2))
loglik = bbr$loglik

loglikPlot(loglik)
}
\seealso{
\code{\link{logLik.bayesbr}},\code{\link{envelope}}
}
