% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.bayesbr.R
\name{model.bayesbr}
\alias{model.bayesbr}
\title{Matrix with All Variables for \code{bayesbr} Objects}
\usage{
\method{model}{bayesbr}(Y,X = NULL,W = NULL,name_y,names_x = NULL,
             names_w = NULL)
}
\arguments{
\item{Y}{A vector containing the model response variable,}

\item{X}{A matrix containing the covariates for theta of the model,}

\item{W}{A matrix containing the covariates of the model for zeta,}

\item{name_y}{The name passed in the call to the \code{\link{bayesbr}} function for the variable response,}

\item{names_x}{The name passed in the call to the \code{\link{bayesbr}} function for the covariates for theta,}

\item{names_w}{The name passed in the call to the \code{\link{bayesbr}} function for covariates for zeta.}
}
\value{
A matrix containing all variables in the model and their names used as column names.
}
\description{
The function receives all variables and their respective names, and concatenates them in a matrix.
}
\seealso{
\code{\link{values}},\code{\link{bayesbr}}
}
