% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_precision.R
\name{summary_precision}
\alias{summary_precision}
\title{Variable Coefficients for Zeta}
\usage{
summary_precision(x,prob=0.95)
}
\arguments{
\item{x}{an object of the class \emph{bayesbr}, containing the list returned from the \code{\link{bayesbr}} function.}

\item{prob}{a probability containing the credibility index for the HPD interval for the coefficients of the covariates.}
}
\value{
A list containing the estimates for the covariables of zeta, this list contains the following items:
\describe{
\item{table}{a table with the means, medians, standard deviations and the Highest Posterior Density (HPD) Interval,}
\item{coeff}{a vector containing the estimated coefficients for the variables.}}
}
\description{
A function that uses the gamma values of the posterior distribution of the model and calculates the estimates for each zeta covariate.
}
\seealso{
\code{\link{summary_mean}},\code{\link{values}},\code{\link{summary.bayesbr}}
}
