% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tau_xi.R
\name{summary_tau_xi}
\alias{summary_tau_xi}
\title{Coefficients for tau_xi}
\usage{
summary_tau_xi(x,prob=0.95)
}
\arguments{
\item{x}{an object of the class \emph{bayesbr}, containing the list returned from the \code{\link{bayesbr}} function.}

\item{prob}{a probability containing the credibility index for the HPD interval for the coefficients of the covariates.}
}
\value{
A list containing the estimates for tau parametrer, this list contains the following items:
\describe{
\item{table}{a table with the means, medians, standard deviations and the Highest Posterior Density (HPD) Interval,}
\item{coeff}{a vector containing the estimated coefficients.}}
}
\description{
A function that uses values of the posterior distribution of the model and calculates the estimates for tau_xi parametrer.
}
\seealso{
\code{\link{summary_xi}},\code{\link{values}},\code{\link{summary_tau_delta}}
}
