% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesbr_app.R
\name{bayesbr_app}
\alias{bayesbr_app}
\alias{bayesbr_app,}
\alias{bayesbr_shiny}
\title{bayesbr Application}
\usage{
bayesbr_app()
}
\description{
A function that runs the shiny application designed for using bayesbr package functions through a visual interface.
}
\details{
See the application manual:
}
\examples{
\donttest{
bayesbr_app()
}
}
\seealso{
\code{\link{bayesbr}}, \code{\link{shiny}}
}
